Cleanup, move mouse/keyboard input code out of emulator.c
