+++ input/input.h
@@ -1,0 +1,2 @@
+int get_mouse_status(char *x, char *y, char *b);
+int kbhit();
